refactor: create LectureListCode enum
