+++ src/components/sections/timetable/lecturedetail/LectureDetailSection.jsx
@@ -21,7 +21,7 @@ import {
 } from '../../../../actions/timetable/timetable';
 
 import { LectureFocusFrom } from '@/shapes/enum';
-import { LectureListCode } from '../../../../reducers/timetable/list';
+import { LectureListCode } from '@/shapes/enum';
 
 import userShape from '../../../../shapes/model/session/UserShape';
 import lectureFocusShape from '../../../../shapes/state/timetable/LectureFocusShape';
